chore(#8): Hall 관련 dto 패키시 위치 수정
