+++ performance/src/main/java/com/ticketing/performance/application/dto/hall/CreateHallResponseDto.java
@@ -1,10 +1,8 @@
-package com.ticketing.performance.application.dto;
+package com.ticketing.performance.application.dto.hall;
 
 import com.ticketing.performance.domain.model.Hall;
-import com.ticketing.performance.domain.model.HallSeat;
 import lombok.*;
 
-import java.util.List;
 import java.util.UUID;
 
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
