+++ performance/src/main/java/com/ticketing/performance/application/dto/hall/HallInfoResponseDto.java
@@ -1,4 +1,4 @@
-package com.ticketing.performance.application.dto;
+package com.ticketing.performance.application.dto.hall;
 
 import com.ticketing.performance.domain.model.Hall;
 import lombok.AllArgsConstructor;
@@ -6,25 +6,30 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
 import java.util.UUID;
 
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
 @Getter
-public class HallListResponseDto {
+@Builder
+public class HallInfoResponseDto {
 
     private UUID hallId;
     private String hallName;
     private String hallAddress;
     private Integer totalSeat;
+    private List<HallSeatInfoResponseDto> seats;
 
-    public static HallListResponseDto of(Hall hall) {
-        return HallListResponseDto.builder()
+    public static HallInfoResponseDto of(Hall hall) {
+        List<HallSeatInfoResponseDto> seats = hall.getHallSeats().stream().map(HallSeatInfoResponseDto::of).toList();
+
+        return HallInfoResponseDto.builder()
                 .hallId(hall.getId())
-                .hallName(hall.getHallName())
                 .hallAddress(hall.getHallAddress())
+                .hallName(hall.getHallName())
                 .totalSeat(hall.getTotalSeat())
+                .seats(seats)
                 .build();
     }
 }
