+++ performance/src/main/java/com/ticketing/performance/application/dto/hall/UpdateHallResponseDto.java
@@ -1,4 +1,4 @@
-package com.ticketing.performance.application.dto;
+package com.ticketing.performance.application.dto.hall;
 
 import com.ticketing.performance.domain.model.Hall;
 import lombok.AllArgsConstructor;
@@ -6,30 +6,26 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
 import java.util.UUID;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Builder
-public class HallInfoResponseDto {
+public class UpdateHallResponseDto {
 
     private UUID hallId;
     private String hallName;
     private String hallAddress;
     private Integer totalSeat;
-    private List<HallSeatInfoResponseDto> seats;
 
-    public static HallInfoResponseDto of(Hall hall) {
-        List<HallSeatInfoResponseDto> seats = hall.getHallSeats().stream().map(HallSeatInfoResponseDto::of).toList();
-
-        return HallInfoResponseDto.builder()
+    public static UpdateHallResponseDto of(Hall hall) {
+        return UpdateHallResponseDto.builder()
                 .hallId(hall.getId())
-                .hallAddress(hall.getHallAddress())
                 .hallName(hall.getHallName())
+                .hallAddress(hall.getHallAddress())
                 .totalSeat(hall.getTotalSeat())
-                .seats(seats)
                 .build();
     }
+
 }
